backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
tiktoken
openai
streamlit
